+++ CHANGES.rst
@@ -22,6 +22,7 @@ Unreleased
 * Add the ``broken`` argument to :meth:`~Reader.get_feeds`,
   which allows getting only feeds that failed / did not fail
   during the last update.
+  Allow filtering feeds by update success in the web application.
   (:issue:`189`)
 * In the ``preview_feed_list`` plugin, add ``<meta name="alternate" ...>``
   to the feed detection heuristics.